+++ app/build.gradle
@@ -18,10 +18,8 @@ repositories {
 dependencies {
     implementation project(':mvc')
 
-    implementation 'org.apache.tomcat.embed:tomcat-embed-core:10.1.25'
-    implementation 'org.apache.tomcat.embed:tomcat-embed-jasper:10.1.25'
     implementation 'ch.qos.logback:logback-classic:1.5.7'
-    implementation 'org.apache.commons:commons-lang3:3.14.0'
+    implementation 'jakarta.servlet:jakarta.servlet-api:6.0.0'
 
     testImplementation 'org.assertj:assertj-core:3.26.0'
     testImplementation 'org.mockito:mockito-core:5.12.0'
