+++ mvc/build.gradle
@@ -21,10 +21,13 @@ dependencies {
     implementation 'jakarta.annotation:jakarta.annotation-api:2.1.1'
     annotationProcessor 'jakarta.annotation:jakarta.annotation-api:2.1.1'
 
+
     implementation 'org.reflections:reflections:0.10.2'
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.1'
     implementation 'ch.qos.logback:logback-classic:1.5.6'
     implementation 'org.apache.commons:commons-lang3:3.14.0'
+    implementation 'org.apache.tomcat.embed:tomcat-embed-core:10.1.25'
+    implementation 'org.apache.tomcat.embed:tomcat-embed-jasper:10.1.25'
 
     testImplementation 'org.assertj:assertj-core:3.26.0'
     testImplementation 'org.mockito:mockito-core:5.12.0'
