+++ mvc/src/main/java/com/interface21/web/TomcatStarter.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.web;
 
 import org.apache.catalina.Context;
 import org.apache.catalina.LifecycleException;
